// alfcomponents/components/simple/alf-button/enums/defaultButtonKeys.interface.ts
/**
 * Keys for default button configurations
 * These are used as identifiers, while the label property contains the display text.
 * Expanded to cover full Bootstrap 5 set.
 */
export enum DefaultButtonKeys {
    Accept = 'ACCEPT',
    Cancel = 'CANCEL',
    Danger = 'DANGER',
    Warning = 'WARNING',
    Info = 'INFO',
    Light = 'LIGHT',
    Back = 'BACK',
    Example = 'EXAMPLE',
    // Geek Variants
    Cyber = 'CYBER',
    Matrix = 'MATRIX',
    Jedi = 'JEDI',
    Sith = 'SITH',
    Fire = 'FIRE',
    Frost = 'FROST',
    Lava = 'LAVA',
    Zen = 'ZEN',
}